# [Bug #20710] Implement global allocatable slots and empty pages